requirements.txt
langchain-core==0.1.22
langchain==0.1.12
langgraph==0.0.25
langsmith==0.0.83
langchain-openai==0.0.8
python-dotenv==1.0.1
openai==1.12.0